Re-opening notatistic forces old window to front.
Don't laugh, but this is shockingly difficult to do in tkInter. First,
you need to continually scan for someone to press the magic keys
anywhere.

The only way to do that is to recognize that the magic keys trigger
Notastic to ope, so you can check for someone to open up Notastic, and
then check for other open Notatistic instances (done via PID file right
now.) If there is one, Notastic needs to  communicate with it somehow
(done by writing contents to the PID file - a slow solution but
definitely way easier than say opening up a socket connection or
something crazy like that. And it's not like you need lightning fast
performance here anyways.  Keep it simple.)

Then, you actually need to:

1) Force app to top.
2) Focus on it.

Force the app to top can only be done with a Toplevel, which root
fortunately is (took me a while to figure that out!) - you do root.lift.

Focusing is another whole bag of worms. You have to deiconify and
iconify the window to get it to focus. I don't know what that means. I
just found it on StackOverflow, and it seems to work alright.

TkInter kinda scares me now. I'm thinking about migrating to another GUI
library, but they all apparently suck - except maybe pyQt? And it sucks
to have to port all that text manip stuff I'm doing :/